Removed the laziness from quick-fill
The quick-fill function was returning a lazy sequence, which was not
realized until it was consumed. Hence if the sequence was never
consumed, the form would not get filled out. Adding a doall to the code
in quick-fill forces the realization of the sequence.

This commit also contains some changes to whitespace.