Relax initialization of desired capabilities
This PR changes the way the DesiredCapabilities object gets
initialized. Instead of using the constructor that uses a map it uses
the no argument constructor and sets the capabilities via the
`setCapability` method.

This is important when running on SauceLabs, since the map constructor
does some additional checks and throws an exception when setting some
properties that are used by SauceLabs. This PR makes it possible to
use the properties that are shown when configuring capabilities via
SauceLabs platform configurator.

https://docs.saucelabs.com/reference/platforms-configurator