Clean up most reflection warnings
Have left some in place for things that I expect to be refactoring
heavily anyway. Window tests also not working now, because there were
some erroneous assumptions in the implementation about the ordering of
windows as they come back from Selenium-WebDriver, so all of the
window-handling code needs to be reapproached.