Almost get exception handling right
Still haven't found the right design for exception catching in the monad
yet. If an exception occurs in the first step in a monadic computation,
then it's not being handled by m-bind yet, so it's "in the open" and
with the current code I don't catch that first exception. This commit
contains the handling code inside m-bind.

For the other case, I either need to go with a separate `drive-...`
macro for each supported monad (and make it special-case the first
underlying binding for `domonad` to catch the exception) or I need to
increase the smarts in `drive` to determine if it should do the
same. I've added `::catch` metadata as a start for a way to capture in
the code which of the monads would need this "catch my first exception
for me" behavior and which would not.